Put tabs component in different file.
